requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
websockets==11.0.3
python-multipart==0.0.6
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.5
fastapi==0.68.0
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib==1.7.4